Added Renderer Component and a View Class, can currently load and render bitmaps. Does not render bitmaps that are not in the view. Views can be instantiated and can even view the same thing at the same time.
Deleted the IGameObjectFactory interface, it was not necessary.

Added a few classes that I've made in the past. They can become very handy.

Also added the SGCL parser.
It is not plugged into the system yet.
I'll do so in the near future.

Added a 3rd party library that SpriteSheet.as, one of my old classes, is dependant on.