Added the Camera component. Now, in the constructor of the View class, a GameObject is instantiated and a Camera component is added to it.
Fixed bugs related to transformations: rotation, scaling and shearing are now correct. Now, the transformation of nested renderers are correct (Renderer of a GameObject that has a parent GameObject).

Added a toString() method for both the GameObject and the Component class.
Added a log() method for both the GameObject and the Component class. It is preferable that we will use log() instead of trace() to trace out stuff. Also, by calling log() without parameters on a Component will make it trace out it's type, properties and variables.
With GameObjects it will make it trace out it's components(type, properties and variables) and what other GameObjects it contains.

Added a smart cache for loading images. Now, images are never loaded more than once.

Added ability to control the transformation offset, smoothing and pixel snapping of a renderer.